requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
python-dotenv==0.19.0
smtplib==0.0.2
